requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
redis==5.1.1
pymongo==4.9.2
langchain==0.3.7
langchain-core==0.3.16
langgraph==0.2.28
chromadb==0.5.15
pydantic==2.9.2
ray==2.50.1
python-dotenv==1.0.1
pypdf==5.0.1
langchain-google-genai==2.0.2
reportlab==4.2.5
markdown==3.7
